+++ model/LogisticRegression.py
@@ -1,42 +1,7 @@
 from sklearn.linear_model import LogisticRegression
 from model.base import BaseModel
-from modelling.data_model import Data
-from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, classification_report
-import numpy as np
-
-# This file already contain the code for implementing LogisticRegression model.
 
 class LogisticRegressionModel(BaseModel):
     def __init__(self) -> None:
         super(LogisticRegressionModel, self).__init__()
         self.model = LogisticRegression(max_iter=1000)
-
-    def train(self, data: Data) -> None:
-        self.model.fit(data.get_X_train(), data.get_type_y_train())
-        # Get predictions on the test set
-        y_pred = self.model.predict(data.get_X_test())
-        print(f"Model predictions: {y_pred}")   
-         
-        # Get the actual test labels
-        y_true = data.get_type_y_test()
-        
-        # Calculate metrics
-        accuracy = accuracy_score(y_true, y_pred)
-        precision = precision_score(y_true, y_pred, average='weighted', zero_division=1)
-        recall = recall_score(y_true, y_pred, average='weighted', zero_division=1)
-        f1 = f1_score(y_true, y_pred, average='weighted', zero_division=1)
-        
-        print(f"Model accuracy: {accuracy:.2f}")
-        print(f"Model precision: {precision:.2f}")
-        print(f"Model recall: {recall:.2f}")
-        print(f"Model F1-score: {f1:.2f}")
-        
-        print("\nClassification Report:")
-        print(classification_report(y_true, y_pred, zero_division=1))
-
-    def predict(self, data: Data):
-        if self.model is None:
-            raise ValueError("The model has not been trained yet. Train the model before calling predict().")
-
-        self.predictions = self.model.predict(data.get_X_test())  # Generate predictions
-        return self.predictions  #returns predictions
